requirements.txt
anthropic>=0.39.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
prompt-toolkit>=3.0.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
